+++ src/bridge/IBridgeSettlerActions.sol
@@ -41,4 +41,7 @@ interface IBridgeSettlerActions {
 
     /// @dev Bridge ERC20 through LayerZeroOFT
     function BRIDGE_ERC20_TO_LAYER_ZERO_OFT(address token, address oft, bytes calldata sendData) external;
+
+    /// @dev Bridge ERC20 through DeBridge
+    function BRIDGE_TO_DEBRIDGE(uint256 globalFee, bytes calldata createOrderData) external;
 }